+++ java/com/turbovnc/vncviewer/F8Menu.java
@@ -1,6 +1,6 @@
 /* Copyright (C) 2002-2005 RealVNC Ltd.  All Rights Reserved.
  * Copyright (C) 2011, 2013 Brian P. Hinz
- * Copyright (C) 2012-2015, 2017-2018, 2020 D. R. Commander.
+ * Copyright (C) 2012-2015, 2017-2018, 2020-2021 D. R. Commander.
  *                                          All Rights Reserved.
  *
  * This is free software; you can redistribute it and/or modify
@@ -70,7 +70,7 @@ public class F8Menu extends JPopupMenu implements ActionListener {
                   KeyEvent.VK_Z);
     showToolbar = new JCheckBoxMenuItem("Show Toolbar   (Ctrl-Alt-Shift-T)");
     showToolbar.setMnemonic(KeyEvent.VK_T);
-    showToolbar.setSelected(cc.showToolbar);
+    showToolbar.setSelected(cc.opts.showToolbar);
     showToolbar.addActionListener(this);
     add(showToolbar);
     tileWindows = addMenuItem("Tile All Viewer Windows   (Ctrl-Alt-Shift-X)",
@@ -148,7 +148,7 @@ public class F8Menu extends JPopupMenu implements ActionListener {
       cc.toggleFullScreen();
     } else if (actionMatch(ev, showToolbar)) {
       cc.toggleToolbar();
-      showToolbar.setSelected(cc.showToolbar);
+      showToolbar.setSelected(cc.opts.showToolbar);
     } else if (actionMatch(ev, defaultSize)) {
       cc.sizeWindow();
       firePopupMenuCanceled();